+++ lib/audio_metadata_reader.dart
@@ -1,6 +1,7 @@
 library audio_metadata_reader;
 
-export "src/parsers/tag_parser.dart" show AudioMetadata, InvalidTag, Picture;
+export "src/parsers/tag_parser.dart" show AudioMetadata, Picture;
 export 'src/metadata/base.dart' show PictureType;
+export 'src/utils/metadata_parser_exception.dart' show MetadataParserException;
 export 'src/parser.dart' show readMetadata;
 export 'src/writer.dart' show writeMetadata;