+++ lib/src/parser.dart
@@ -1,12 +1,13 @@
 import 'dart:io';
 
+import 'package:audio_metadata_reader/audio_metadata_reader.dart';
 import 'package:audio_metadata_reader/src/metadata/mp3_metadata.dart';
 import 'package:audio_metadata_reader/src/metadata/mp4_metadata.dart';
 import 'package:audio_metadata_reader/src/metadata/vorbis_metadata.dart';
+import 'package:audio_metadata_reader/src/parsers/id3v1.dart';
 import 'package:audio_metadata_reader/src/parsers/id3v2.dart';
 import 'package:audio_metadata_reader/src/parsers/mp4.dart';
 import 'package:audio_metadata_reader/src/parsers/ogg.dart';
-import 'package:audio_metadata_reader/src/parsers/tag_parser.dart';
 import 'package:audio_metadata_reader/src/parsers/flac.dart';
 
 /// Parse the metadata of a file.
@@ -131,9 +132,42 @@ AudioMetadata readMetadata(File track, {bool getImage = false}) {
 
       return newMetadata;
     }
-  } catch (e) {
-    return InvalidTag(file: File(""));
+    if (ID3v2Parser.isID3v1(reader)) {
+      final mp3Metadata = ID3v1Parser().parse(reader) as Mp3Metadata;
+
+      final a = AudioMetadata(
+        file: track,
+        album: mp3Metadata.album,
+        artist: mp3Metadata.bandOrOrchestra ??
+            mp3Metadata.originalArtist ??
+            mp3Metadata.leadPerformer,
+        bitrate: mp3Metadata.bitrate,
+        duration: mp3Metadata.duration,
+        language: mp3Metadata.languages,
+        lyrics: mp3Metadata.lyric,
+        sampleRate: mp3Metadata.samplerate,
+        title: mp3Metadata.songName,
+        totalDisc: mp3Metadata.totalDics,
+        trackNumber: mp3Metadata.trackNumber,
+        trackTotal: mp3Metadata.trackTotal,
+        year:
+            DateTime(mp3Metadata.originalReleaseYear ?? mp3Metadata.year ?? 0),
+        discNumber: mp3Metadata.discNumber,
+      );
+
+      a.pictures = mp3Metadata.pictures;
+      a.genres = mp3Metadata.genres;
+
+      return a;
+    }
+  } catch (e, trace) {
+    print(trace);
+    throw MetadataParserException(track: track, message: e.toString());
   }
 
-  return InvalidTag(file: File(""));
+  throw MetadataParserException(
+    track: track,
+    message:
+        "No available parser for this file. Please raise an issue in Github",
+  );
 }