+++ lib/src/parsers/tag_parser.dart
@@ -65,10 +65,6 @@ class AudioMetadata {
   }
 }
 
-class InvalidTag extends AudioMetadata {
-  InvalidTag({required super.file});
-}
-
 abstract class TagParser {
   final bool fetchImage;