+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 ## 0.0.7
 
-- BREAKING CHANGE : the API is now synchronous
-- Performance improvement : 70% faster by using sync IO operations and a File Buffer to read files
+### BREAKING CHANGE
+
+- `readMetadata` is now synchronous. You may remove some `Future.wait()`
+- `InvalidTag` class has beem removed and replace with a real exception `MetadataParserException`
+
+## Other
+
+- Performance improvement : **70%** faster by using sync IO operations and a File Buffer to read files
+- Support `ID3v1`
 
 ## 0.0.6